+++ src/leapmotor_api/const.py
@@ -20,6 +20,7 @@ KNOWN_ACCOUNT_P12_PASSWORDS: tuple[str, ...] = ()
 # Remote-control action identifiers.
 REMOTE_CTL_LOCK = "lock"
 REMOTE_CTL_UNLOCK = "unlock"
+REMOTE_CTL_UNLOCK_CHARGER = "unlock_charger"
 REMOTE_CTL_TRUNK = "trunk"
 REMOTE_CTL_TRUNK_OPEN = "trunk_open"
 REMOTE_CTL_TRUNK_CLOSE = "trunk_close"